# outputs/openapi_schema.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "TRACE Extreme-Weather Alert API",
    "description": "Research-prototype API built on frozen TRACE evidence. Severity tiers are prototype heuristics, not official meteorological warnings or calibrated probabilities. All outputs require human review and cannot self-authorize public warning release.",
    "version": "1.8.0-prototype"
  },
  "paths": {
...
    "/api/v1/tracks/build": {
      "post": {
        "summary": "Build Track",
        "operationId": "build_track_api_v1_tracks_build_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TrackRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Build Track Api V1 Tracks Build Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/provider-ingest/validate": {
      "post": {
        "summary": "Provider Ingest Validate",
        "operationId": "provider_ingest_validate_api_v1_provider_ingest_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProviderMetadataRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Provider Ingest Validate Api V1 Provider Ingest Validate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/stream/replay": {
      "post": {
        "summary": "Controlled Stream Replay",
        "operationId": "controlled_stream_replay_api_v1_stream_replay_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ControlledReplayRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Controlled Stream Replay Api V1 Stream Replay Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/alerts/{lead_hours}": {
      "get": {
        "summary": "Alert",
        "operationId": "alert_api_v1_alerts__lead_hours__get",
        "parameters": [
          {
            "name": "lead_hours",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lead Hours"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Alert Api V1 Alerts  Lead Hours  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/refinement-request/{lead_hours}": {
      "get": {
        "summary": "Refinement Request",
        "operationId": "refinement_request_api_v1_refinement_request__lead_hours__get",
        "parameters": [
          {
            "name": "lead_hours",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lead Hours"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Refinement Request Api V1 Refinement Request  Lead Hours  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/impact-object/{lead_hours}": {
      "get": {
        "summary": "Impact Object",
        "operationId": "impact_object_api_v1_impact_object__lead_hours__get",
        "parameters": [
          {
            "name": "lead_hours",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Lead Hours"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Impact Object Api V1 Impact Object  Lead Hours  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/alerts/evaluate": {
      "post": {
        "summary": "Evaluate Alert",
        "operationId": "evaluate_alert_api_v1_alerts_evaluate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "type": "object",
                  "title": "Response Evaluate Alert Api V1 Alerts Evaluate Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/severity": {
      "post": {
        "summary": "Severity",
        "operationId": "severity_api_v1_severity_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvidenceInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeverityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AlertRequest": {
        "properties": {
          "event_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Event Id"
          },
          "valid_time": {
            "type": "string",
            "format": "date-time",
            "title": "Valid Time"
          },
          "forecast_lead_hours": {
            "type": "integer",
            "maximum": 240.0,
            "minimum": 72.0,
            "title": "Forecast Lead Hours",
            "description": "Medium-range forecast lead, 3 to 10 days."
          },
          "core_lat": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Core Lat"
          },
          "core_lon": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Core Lon"
          },
          "macro_footprint_area_km2": {
            "type": "number",
            "maximum": 510100000.0,
            "exclusiveMinimum": 0.0,
            "title": "Macro Footprint Area Km2"
          },
          "macro_footprint_geojson": {
            "additionalProperties": true,
            "type": "object",
            "title": "Macro Footprint Geojson"
          },
          "evidence": {
            "$ref": "#/components/schemas/EvidenceInput"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "event_id",
          "valid_time",
          "forecast_lead_hours",
          "core_lat",
          "core_lon",
          "macro_footprint_area_km2",
          "macro_footprint_geojson",
          "evidence"
        ],
        "title": "AlertRequest"
      },
      "ControlledReplayRequest": {
        "properties": {
          "records": {
            "items": {
              "$ref": "#/components/schemas/ReplayRecord"
            },
            "type": "array",
            "maxItems": 100,
            "minItems": 1,
            "title": "Records"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "records"
        ],
        "title": "ControlledReplayRequest"
      },
      "EvidenceInput": {
        "properties": {
          "max_ensemble_mean_wind_ms": {
            "type": "number",
            "minimum": 0.0,
            "title": "Max Ensemble Mean Wind Ms",
            "description": "Maximum ensemble-mean wind speed inside the candidate footprint."
          },
          "max_efi": {
            "type": "number",
            "maximum": 1.0,
            "minimum": -1.0,
            "title": "Max Efi",
            "description": "Extreme Forecast Index. Positive extremes approach +1."
          },
          "max_member_support_fraction": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Max Member Support Fraction"
          },
          "mean_seed_agreement": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Mean Seed Agreement"
          },
          "max_sot": {
            "type": "number",
            "title": "Max Sot",
            "description": "Shift of Tails evidence. Positive values corroborate tail shift."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "max_ensemble_mean_wind_ms",
          "max_efi",
          "max_member_support_fraction",
          "mean_seed_agreement",
          "max_sot"
        ],
        "title": "EvidenceInput"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ProviderMetadataRequest": {
        "properties": {
          "provider_name": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "title": "Provider Name"
          },
          "dimensions": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Dimensions"
          },
          "variables": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              }
            ],
            "title": "Variables"
          },
          "coordinates": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              }
            ],
            "title": "Coordinates"
          },
          "initialization_time": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Initialization Time"
          },
          "valid_times": {
            "items": {
              "type": "string",
              "format": "date-time"
            },
            "type": "array",
            "title": "Valid Times"
          },
          "lead_hours": {
            "items": {
              "type": "number"
            },
            "type": "array",
            "title": "Lead Hours"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "provider_name",
          "dimensions",
          "variables"
        ],
        "title": "ProviderMetadataRequest"
      },
      "ReplayRecord": {
        "properties": {
          "event_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Event Id"
          },
          "forecast_lead_hours": {
            "type": "integer",
            "maximum": 240.0,
            "minimum": 72.0,
            "title": "Forecast Lead Hours"
          },
          "severity": {
            "type": "string",
            "enum": [
              "LOW",
              "MODERATE",
              "SEVERE"
            ],
            "title": "Severity"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "event_id",
          "forecast_lead_hours",
          "severity"
        ],
        "title": "ReplayRecord"
      },
      "SeverityResponse": {
        "properties": {
          "schema": {
            "type": "string",
            "const": "trace-severity-response/1.5",
            "title": "Schema",
            "default": "trace-severity-response/1.5"
          },
          "api_version": {
            "type": "string",
            "title": "Api Version",
            "default": "1.8.0-prototype"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version",
            "default": "trace-wind-spherical-gnn-bulbul-frozen-5seed"
          },
          "evidence_version": {
            "type": "string",
            "title": "Evidence Version",
            "default": "bulbul2019-alert-evidence-v1-fani-temporal-v1-stage2-controlled-hrrr-v3-governance-v1"
          },
          "prototype_alert": {
            "type": "boolean",
            "title": "Prototype Alert",
            "default": true
          },
          "official_warning": {
            "type": "boolean",
            "title": "Official Warning",
            "default": false
          },
          "calibrated_probability": {
            "type": "boolean",
            "title": "Calibrated Probability",
            "default": false
          },
          "severity": {
            "type": "string",
            "enum": [
              "LOW",
              "MODERATE",
              "SEVERE"
            ],
            "title": "Severity"
          },
          "rule": {
            "type": "string",
            "title": "Rule"
          },
          "rule_provenance": {
            "type": "string",
            "title": "Rule Provenance"
          },
          "thresholds": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Thresholds"
          },
          "explanation": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Explanation"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "severity",
          "rule",
          "rule_provenance",
          "thresholds",
          "explanation"
        ],
        "title": "SeverityResponse"
      },
      "TrackFrame": {
        "properties": {
          "valid_time": {
            "type": "string",
            "format": "date-time",
            "title": "Valid Time"
          },
          "forecast_lead_hours": {
            "type": "integer",
            "maximum": 384.0,
            "minimum": 0.0,
            "title": "Forecast Lead Hours"
          },
          "core_lat": {
            "type": "number",
            "maximum": 90.0,
            "minimum": -90.0,
            "title": "Core Lat"
          },
          "core_lon": {
            "type": "number",
            "maximum": 180.0,
            "minimum": -180.0,
            "title": "Core Lon"
          },
          "macro_footprint_geojson": {
            "additionalProperties": true,
            "type": "object",
            "title": "Macro Footprint Geojson"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "valid_time",
          "forecast_lead_hours",
          "core_lat",
          "core_lon",
          "macro_footprint_geojson"
        ],
        "title": "TrackFrame"
      },
      "TrackRequest": {
        "properties": {
          "event_id": {
            "type": "string",
            "maxLength": 128,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_.:-]+$",
            "title": "Event Id"
          },
          "frames": {
            "items": {
              "$ref": "#/components/schemas/TrackFrame"
            },
            "type": "array",
            "maxItems": 64,
            "minItems": 2,
            "title": "Frames"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "event_id",
          "frames"
        ],
        "title": "TrackRequest"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}